마커 안보이는 오류 수정 user api response 타입 수정

+++ app/src/main/java/com/mommydndn/app/data/api/model/response/JobOfferResponse.kt
@@ -100,5 +100,5 @@ data class LatestReview(
     @SerialName("nickname")
     val nickname: String,
     @SerialName("rate")
-    val rate: Int
+    val rate: Double
 )
